% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMCM-package.R
\docType{package}
\name{GMCM-package}
\alias{GMCM-package}
\alias{GMCM}
\title{Fast optimization of Gaussian Mixture Copula Models}
\description{
Gaussian mixture copula models (GMCM) are a flexible class of statistical
models which can be used for unsupervised clustering, meta analysis, and
many other things. In meta analysis, GMCMs can be used to
quantify and identify which features which have been reproduced across
multiple experiments. This package provides a fast and general
implementation of GMCM cluster analysis and serves as an improvement and
extension of the features available in the \code{idr} package.
}
\details{
If the meta analysis of Li et al. (2011) is to be performed, the
  function is used to identify the maximum
  likelihood estimate of the special Gaussian mixture copula model (GMCM)
  defined by Li et al. (2011). The function
  computes the local and adjusted Irreproducible Discovery Rates defined
  by Li et al. (2011) to determine the level of reproducibility.

  Tewari et. al. (2011) proposed using GMCMs as an general unsupervised
  clustering tool. If such a general unsupervised clustering is needed, like
  above, the function  computes the maximum
  likelihood estimate of the general GMCM. The function
  is used to estimate the class membership
  probabilities of each observation.

  provide easy simulation from the GMCMs.
}
\references{
Anders Ellern Bilgrau, Poul Svante Eriksen, Jakob Gulddahl Rasmussen,
  Hans Erik Johnsen, Karen Dybkaer, Martin Boegsted (2016). GMCM:
  Unsupervised Clustering and Meta-Analysis Using Gaussian Mixture Copula
  Models. Journal of Statistical Software, 70(2), 1-23.
  doi:10.18637/jss.v070.i02

  Li, Q., Brown, J. B. J. B., Huang, H., & Bickel, P. J. (2011).
  Measuring reproducibility of high-throughput experiments. The Annals of
  Applied Statistics, 5(3), 1752-1779. doi:10.1214/11-AOAS466

  Tewari, A., Giering, M. J., & Raghunathan, A. (2011). Parametric
  Characterization of Multimodal Distributions with Non-gaussian Modes.
  2011 IEEE 11th International Conference on Data Mining Workshops,
  286-292. doi:10.1109/ICDMW.2011.135

  Package by Li et. al. (2011): \code{\link[idr:idr-package]{idr}}.
}
\author{
Anders Ellern Bilgrau,
  Martin Boegsted,
  Poul Svante Eriksen

  Maintainer: Anders Ellern Bilgrau <anders.ellern.bilgrau@gmail.com>
}
